requirements.txt
asgiref==3.10.0
Django==5.2.8
django-debug-toolbar==6.1.0
django-rest-framework==0.1.0
sqlparse==0.5.3
tzdata==2025.2
